Support for "gain Bounty" abilities (#161)
Discovered a timing bug in the original Bounty PR that prevented "gain
Bounty" abilities on upgrades from working correctly because the defeat
of the attached upgrade was removing the ability before the defeat of
the unit could trigger it.

Reordered the timing of ability and trigger registration /
de-registration to not happen until after all events in the window have
finished resolving. This required some tweaks in various places that
were making assumptions about game state and card location.

Additionally, noticed some opportunities to improve the target selection
logic to cut back on unnecessary prompts, particularly when related to
exhausting / readying cards. In some cases prompts were being emitted to
ready / exhaust cards when all available targets were already in that
state, in which case the user's selection is pointless. Improved the
targeting logic and several related test cases.

Fixed all of the above and added the Wanted and Top Target cards.