Move card target condition check before system check (#179)
Rearranged things in CardTargetResolver so that that the condition
filter is checked before the system's hasLegalTarget is invoked. This is
because the condition filter might be able to prevent the system from
trying to properties are cards that are in an invalid state (e.g.,
attempting to read the damage off of a card in hand).