Add shortcut in ability resolver for no effective targets (#232)
We found that in some cases, marking a triggered ability optional would
case a prompt to be displayed for activating the ability even if there
were no "effective" targets (e.g., an exhaust ability trying to activate
when all units are exhausted). Added a shortcut for this, tested in a
different branch.