Fix uniqueness check when multiple unique units enter play
Previously, we were assuming that only one copy of a unique card could be moved in any given window, but this is not the case when captured cards are rescued. So we ended up with multiple stacked prompts that could never resolve. Now we account for this by only prompting once per copy of a unique card that enters play, and allowing multiple selection if there are more then two in play.