Game and other state, and unit test improvements (#1035)
### Highlights
* A number of Game.js fields moved into a state object
* Player Decklist converted to use GameObjectRef and moved into state
* Card.parentCard converted to GameObjectRef and moved into state
* Fixed, then disabled the "filter card from list" in Game.js, it wasn't
doing anything before so it's disabled for now.
* Converted undo test to a custom global uit function
* Added a JS private #experimental object in the Game for flagging the
system to use undo logic. That way I can enable it in tests now and in
the future for local builds only.

### Code I'm Not So Proud Of
Soooo I made some methods public in ActionPhase to make clearing the
ActionPhase possible. It's a little horrifying but it did work for
tests. And the way I clear the pipeline also was quick and dirty.

### Testing Notes
the new uit function only works with tests that do the setup in a
beforeEach. In part because the rollback only is working for the
ActionPhase, but also because if the setup is within an _it_ call, it'll
bump into the rollback which happens before the test runs and can cause
complications. So, the uit method will simply treat the test as a normal
test if the setup call hasn't happened yet.

To manually run all tests in undo, simply put `global.it = global.uit;`
after the uit definition. uit will have false positives, but the
failures we do run into will help us hunt down more issues with undo.

---------

Co-authored-by: AMMayberry1 <ammayberry1@gmail.com>