Connecting to the DYNAMODB Successful! (#847)
- Added service for the DynamoDB with basic add,edit,query,delete
functions.
- Updated the environmental variables
- added a test POST call for the FE to contact the dynamoDB
- added 2 dependencies for the DB
- to be able to connect to the DB you need to add three fields in the
local .env file.
AWS_REGION
AWS_ACCESS_KEY_ID
AWS_SECRET_ACCESS_KEY

for the values contact an administrator.

Additionally here is the link to the document to set up a testing local
database

https://docs.google.com/document/d/1juocZ5aC-3BGXeOw2W7EjHfEeKXvHCNGfwIoPyUwvFY/edit?usp=sharing

for testing the
NEXTAUTH_SECRET
env variable needs to be set to the same value as on the FE.

---------

Co-authored-by: AMMayberry1 <ammayberry1@gmail.com>