Added screen resolution and viewport to the bug report entries (#1183)
Updated the bug report to take in the screen resolution and viewport for
helping debug FE issues. Helpful to see what sort of device "roughly"
they are using when complaining about UI features. Strictly captures the
physical screen resolution and then viewport of the game within the
browser

Here is a new bug report from an iphone in landscape

![image](https://github.com/user-attachments/assets/35d0a8a5-835a-4022-bd5c-3b51b6ca0f04)

Here it is in portrait mode

![image](https://github.com/user-attachments/assets/9b54b77f-f6b0-4af2-bafd-08317a57f0bc)

Here is the FE sending the old message w/o screen resolution and
viewport set -- proving its still backward compatible and can read the
old messages

![image](https://github.com/user-attachments/assets/36354322-b68d-4cc4-89e8-8e4f9857db42)