Better profanity filter for username (#1317)
Added the `obscenity` repo which is a more powerful toolkit for checking
for profanity in text. Imported their english default dataset and
tweaked it a bit for our purposes so we could improve our username
filtering (this will also serve well for general chat filtering).

Included a list of common hate group names as well to filter on.