Fix issue with stolen units and direct overwhelm damage (#1321)
Fixes #1320 

Maybe there is a better way to solve this. I was hoping to use
`lastKnownInformation`, but in the `AttackFlow`, I was unable to find a
way to reference the defeat event for the unit that gets defeated before
combat damage is resolved.

Open to feedback if there's a better way to do this!