Refactor player lasting effect (#1342)
There where a few issues with player lasting effects:
- Persistent effects weren't working correctly.
- Some effects were using the card's controller instead of the player
that triggered the effect (e.g. Bendu) so those cards weren't working
correctly if they changed controller.
- The chat logs didn't describe the effects like what we already had for
card lasting effects.

Now `PlayerLastingEffectSystem` accepts a `target` (instead of
`targetController`) that is the player triggering the ability by default
instead of being the controller of the source card. I've also moved the
code to generate chat message to an helper so that it can be shared
between `CardLastingEffectSystem` and `PlayerLastingEffectSystem`.

Co-authored-by: AMMayberry1 <ammayberry1@gmail.com>