Strip out unneeded data from gamestate updates (#1448)
Our gamestate updates from the BE were massive and contained a huge
amount of unused or redundant data. Stripped out the unused fields to
reduce the size to ~1/3 of what it was.

Also consolidated the chat message format a bit to reduce the number of
string pieces being sent.

Corresponding FE PR:
https://github.com/SWU-Karabast/forceteki-client/pull/399