Connection improvements during matchmaking countdown (#1470)
Few small tweaks to patch over some potential problems that might be
causing the queueing issues for people on unstable internet connections.

I suspect that what's going on is that if any disconnects happen during
the handoff from the queue to the lobby, our user connection flag gets
messed up and erroneously marks the user as disconnected. Then, they
don't receive lobby state updates, so the FE never updates and the
server thinks they left.

The changes I made to try and smooth this over are:
- Changed the state-send methods to check the underlying socketio
connection flag, as that is much more reliable. If this seems to work
better then we can plan to move off of our manual connection flag
completely.
- Added a callback on the state-send function that will mark the user's
status as connected if the gamestate or lobbystate message is
acknowledged by the FE
- During the match countdown, lobby state updates will be forced to send
even if we think the user is disconnected. This should increase the
likelihood of the FE actually receiving an update in time.
- Added logic to mark the user as connected if any message is received
from the FE, in case there are other scenarios where the user is
incorrectly getting marked as disconnected and staying that way