Fix cards that search entire deck (#1492)
We have a UI issue currently where the card display prompt can't show
the entire deck since it doesn't have a scroll option, so it can
overwhelm the screen.

Made some adjustments to the cards that need to search the whole deck to
make them at least playable for now:

- Bounty Posting: now will only show the selectable cards from the deck
instead of the whole deck
- Annihilator: already had a similar workaround to Bounty Posting, just
made it _slightly_ less incorrect in that the player can now choose how
many copies to remove from the deck
- Search Your Feelings: now uses a dropdown list of card names instead
the card display. there's a little bit of a rules issue where right now
we can't offer the option to choose no card, since the dropdown UI
doesn't have support for extra buttons. Will work on this