Undo: Ongoing Effect Conversion (#1497)
### Ongoing Effect GO conversion
I converted Ongoing Effect to GOs and that had a trickledown effect to
the OngoingEffectImpl, ValueWrapper, Ability, and AbilityLimit all
needing to be converted to GOs. Some of these don't actually have state
themselves, but they are referenced by other state dynamically, so they
have to be GOs.

That's more GOs than I'd like but to improve it would require doing a
big rework of OngoingEffectImpl and ValueWrapper and that's too much for
this PR alone.

### Ability and AbilityLimit
Because these are all used in the AbilityHelper shortcuts but now
require the game itself, I had to add a bunch of ways to pass in game
when they are constructed.

Converted a lot of the non-state ability arrays into state finally.

### AbilityHelper is now an instance
Because AbilityLimits need a game reference now, I needed to make the
AbilityHelper to hold a reference to game. That is a problem because
it's possible that multiple game instances are running in the same
process, and that would break the static instance. So, it's now an
instance of it's own on the root game and is passed to all of the
different card setups. Hence why there are 1000+ file changes.

There is a StaticAbilityHelper but this was just a shortcut as a couple
of classes in places didn't have easy access to game, and only needed a
single function. Best to avoid this when building cards as it's possible
more of these helpers will need access to game later on.

### And more
Lots of other small changes, probably lost track of some of them, but
it's all likely related to the GO conversions.