Temporary fix for unique upgrades with When Played abilities
When a second copy of a unique upgrade is played, the user must choose to defeat one.
If the user chooses to defeat the newly played copy, we incorrectly fail to trigger any
When Played abilities on the upgrade (Luke's Lightsaber, Mandalorian's Rifle, etc), because
the ability cannot access the LKI to determine if it was attached to the named unit.

This commit temprarily fixes the issue in the case where the new copy is played on the same
unit as the old copy. For this case, we will automatically defeat the oldest copy, allowing
the new copy to resolve its When Played ability.