Experimental: Undo Decorators (#1552)
# Undo Decorators
I recently found JS decorators, which TS 5.0+ implements, and on top of
discussions about the inefficiency of arrays in the state object, I
decided to give this a spin. See the OngoingEffectEngine changes below
for a small example, and GameObjectUtils.ts for the new decorators. It's
fully passing tests.

With this, we can use decorator functions to mark classes and specific
fields for undo rather than having to create the state object field and
the getters/setters manually. The undo decorator overrides get/set on
the auto-accessor and writes data to the state object instead.
`@undoState` is for simple state and is just a pass through to the state
object. `@undoArray` and `@undoObject` are for working with GameObjects
and has additional logic to keep the original references cached, which
can eliminate a bunch of the garbage generated on every get call to
these types of state.

The decoration process makes some changes to the accessor fields
(undoState, undoArray, undoObject) and registers the class as something
that can be undone (registerState) under the hood. It additionally
tracks each decorated accessor by prototype so that branching
inheritance can still undo.

#### Upsides
* Simplifies most state fields to a single decorator and accessor field
`@undoState accessor title: string = ''`, hiding the more complex logic
away, and makes it far easier to add new state, reducing maintenance
cost and making undo more accessible to devs less knowledgeable on how
it works.
* It can cache GameObject's and properly store their GameObjectRefs to
state when they are modified. This cuts down on array map calls and
`this.game.getFromRef()` significantly, saving CPU and garbage objects.
* In theory we could altogether remove GameObjectRefs with this, cutting
out a lot of extra objects.
* The functionality of the decorators could dramatically change w/o
interfering with ongoing work elsewhere.
* Making a decorator more efficient makes anything that uses that
decorator more efficient.
* This can theoretically be used to replace `structuredClone` if we were
to change all state over.

#### Downsides
* There is a lot of "code magic", which I know can be very confusing
when first encountered, and working on the decorators directly is gonna
be harder.
* I think the improved maintenance cost of making it easier for non-undo
devs to create new state and simplifying state as a whole is worth it,
but decorator logic is complex undoubtedly.
* I'll probably still have a way to flag a specific class as "manual"
and it'll just work the way it does now with structuredClone, just in
case.
* This moves a lot of the heavy lifting into the copy function by making
it traverse down the prototype chain looking for metadata to tell it
what to copy; even if there's nothing in a step of the prototype chain
it has to walk down the whole thing and ask, and boy was that a headache
just to understand to begin with.
* Not the most efficient, but this would only happen in rollback so
that's a small price to pay for cutting out all of the extra get calls.
It can definitely be optimized further, and it still may actually still
be faster than structuredClone, which is not fast, but that needs
testing.

#### Other
The declare changes can be ignored. I was testing ES2022 and apparently
there are some [breaking TS
changes](https://github.com/microsoft/TypeScript/issues/56588), one of
which pertains to using override to re-type fields instead of declare. I
rolled the ES version back but forgot to remove those, however it has no
functional difference.