Fix some undo failure cases (#1567)
Fixed a handful of test cases that were failing undo. Most of them were
issues with the test harness, except one.

Some tests involving Bravado were failing because its cost adjuster was
having issues with GO registration. What was happening was that the
`afterSetAllState()` call for `OngoingEffectEngine` was causing some new
GOs to be created, but the GO index property hadn't been reset yet so
they were receiving an index that was too high and then getting
overwritten later.

Reordered the operations in the rollback to avoid this, which caused an
issue with `OngoingEffectEngine` since it was trying to reference GOs
that had already been removed. This was a bad pattern anyway, so I did a
refactor to get around that and convert the `CustomDurationEvents` list
to be a `state` property so that it could just be directly managed via
undo and not need a lot of manual work at the `afterSetAllState()` call.