Automatically clean up GameObjects that aren't referenced (#1574)
I caught that we were holding thousands of PlayerAction game objects
that were generated as part of evaluating potential available actions
for cards. To alleviate this, I added a way to determine whether a
GameObject has ever had a ref created for it. Any GO for which this is
_not_ true at snapshot time will be removed from the list so it can be
GCd.

Special provision had to be made for certain types of objects that are
typically held as "real" references and a `GameObjectRef` is never
created for them. In this case, there is an `alwaysTrackState` property
getter that can be overridden. It's not the most elegant solution but it
will get us by until we can finish the decorator work.

All regular (non-undo) tests are passing but it is possible this
introduced some regressions in the undo work. I will continue
investigating and do a fast-follow PR if so.