Feature: State Watcher Rework (#1590)
Reworked the different state watcher derived classes so there is only
one instance of each class per game. Reduced the functionality of the
StateWatcherRegistrar to simply keep references to the instances and
moved the state back into the individual State Watchers.

For Undo, I added a unregister method for cleaning up the events. I also
added the @undoMap() decorator.

Added more asserts that can take a function in the case where the
message requires an array mapping.

---------

Co-authored-by: AMMayberry1 <ammayberry1@gmail.com>