Fix capture-related exception causing server issues (#1613)
Found that one cause of some of the game crashes is that when an enemy
unit has become captured by another enemy unit, the attack logic handles
it incorrectly and tries to create a hypothetical attack targeting the
captured unit. This causes a targeting exception, which gets hit
repeatedly very fast and exceeds the server error limit.