Prevent transient GOs from being registered as game state (#1614)
Our GC cost is still very high even after cutting back on GOs that are
registered for the entire duration of the game. My guess is that those
GOs are being created and held for a period of time so that they are
being promoted to "old space" in V8, meaning that they still incur a
high GC to traverse until they are finally removed from the map and
released.

This change makes it so that we can declare in advance if a GO is known
to be transient and then prevent it from being registered in the state
map. This should ideally prevent the object from being held longer than
necessary and therefore promoted to old space.