Feature: Zone Undo Decorator Conversion (#1594)
Converted all of the Zone classes to use decorators. 

Added some additional functionality to @undoArray to optionally allow
for the use of in-place modification methods. The default mode is
"readonly" which is what it did previously. The new in-place method is
more complex due to adding a couple of extra layers, but with in-place
mutation it cuts down on additional array creations and all but makes it
work as before. The only function disallowed is directly modifying
indexes, and we don't do that anywhere at the moment. For now, there is
just a safety check and an error message.

Included a little more code for @registerState to prepare to move away
from structuredClone.