Undo: inspect tests and experimental optimization (#1619)
UndoArray filter and map were returning new UndoArray instances. I've
overloaded it to return the results as a plain array to see if chained
maps and filters with UndoArray a part of the problem are.

Added "test-inspect", works with the chrome/edge node debugger so we can
do limited memory/performance profiling from the local machine.

---------

Co-authored-by: AMMayberry1 <ammayberry1@gmail.com>