Handle serialization errors better (#1777)
During undo testing we found that if an undo failure brings the game
into an unserializable state, there's not really much reporting that we
get and the player just sees the screen freeze.

Added some logic to safely serialize and generate a report including as
much as game state as possible on failure. Right now the players will
just receive a blank screen indicating a disconnect, which is not ideal
but at least communicates more clearly that something broke.