Add log if user's authenticated status changes mid-game (#1797)
Added a warning log if a user's status flips between authenticated
(logged in) and not at any point during an ongoing game. This should
help us pinpoint the cause of the issue where chat randomly becomes
enabled or disabled when it shouldn't.

Also bumped card data version as I forgot to do it for the mock data
updates.