Fix issues with repeated triggers in the same window (#1920)
The logic for handling repeated triggers in the same window was assuming
that there were no other abilities triggered in the same window. This
was causing issues in e.g. the scenario where indirect damage is dealt
to multiple units when the player has both the Boba leader and General
Pryde at the same time.

Cleaned up the logic and added tests to confirm that the behavior is now
correct.