Add Vuutun Palaa and advanced cost adjuster interaction logic (#1936)
Added Vuutun Palaa and the logic to handle complex interactions between
Exploit and other cost adjusters which could be impacted by the choice
of targets at the exploit stage.

This solution is very complicated but it mostly involves adding ways for
a cost adjuster at one stage to be able to check what adjustments are
available above and below it so that it can determine which options are
available in terms of target selection. At the "can pay" step, this
means adding an "opportunity cost" concept which lets later stages
inform the Exploit stage of what adjustments would be lost by exploiting
a given unit.

At the trigger stage it's a bit more complicated as we need to be able
to dynamically determine the minimum set of legal targets based on what
the player has already selected. E.g., in some cases it may not be legal
to exploit Vuutun Palaa itself and still be able to pay the full cost,
so in those situations it must not be selectable. Or it could be the
case that selecting it is fine, but then more units must be selected for
Exploit to make up the difference.

This logic was consolidated into a common TargetedCostAdjuster class
from which the exploit and VP adjusters inherit. Also added a lot of
logic to correctly handle the Exploit + Starhawk case, which we weren't
doing correctly before.