Clean up lobby user data tracking (#1962)
We were previously using dynamically generated `PlayerDetails` objects
in the Lobby to track things like user authenticated status and deck
metadata. This was convoluted and causing errors, and since we will be
hitting this code path a lot for Bo3 work it was time to do some
cleanup.

Consolidated these details on the `Player` object that lives inside of
`Game` so that it persists even after a user leaves. The deck data was
all migrated into the `Deck` class which was cleaned up and consolidated
with clearer type signatures. Also cleaned up the requeue flow and the
typing for those parameters.