Cache customization and admin data to reduce DB load (#1966)
We were doing a DB operation for every call to check admin status or to
fetch all customizations (which happens _very_ frequently in the current
design).

Added a simple cache class TimedCache which holds a single value and
refreshes it on a timer, which we use to cache these values.