Introduce Matchmaking Rules (#1968)
## Description

This PR introduces Matchmaking Rules to the `QueueHandler`, starting
with a single `RematchCooldownRule`, which prevents players from being
matched after a specified cooldown period if they faced each other in
their previous match.

The interface makes the feature extensible, so we can easily add more
matchmaking rules in the future. For example, if the players have a
block-list, we could create a new Matchmaking Rule that prevents players
from matching against somebody in their blocklist.

---------

Co-authored-by: AMMayberry1 <ammayberry1@gmail.com>