Fix for Cobb Vanth Allowing Free Play From Hand (#1996)
Fix for the bug reported below. Cobb should now only allow card to be
played for free from discard. If the card is put in hand, it can't be
played for free.

Also put in a test to validate that if the free card is defeated, it
can't be played again for free from discard.


https://discord.com/channels/1220057752961814568/1360259166618259727/1449457181459087394

---------

Co-authored-by: BWohldmann <brian.wohldmann@gmail.comm>